expense-tracking-project/requirements.txt
streamlit==1.51.0
pandas==2.3.3
pytest==9.0.1
fastapi==0.122.0
pydantic==2.12.4
uvicorn==0.38.0
mysql-connector-python==9.5.0
requests==2.32.4
